Implement ability to prefix a path with file or json to include that file.
Instead of writing out the request or response body inline you can now use the
file or json prefix and provide a path to a file you wish to use as the body.

If you use the json prefix the file extension is not required and if the
contents of the file are validated as JSON it will not be re-encoded.